Extract queued texture preparation
